backend/CE/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
bnlearn==0.8.5
cdt==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colourmap==1.1.16
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
d3blocks==1.4.9
d3graph==2.5.0
datazets==0.1.9
decorator==5.1.1
df2onehot==1.0.7
distro==1.9.0
elasticgraph==0.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
funcsigs==1.0.2
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
igraph==0.11.4
ipython==8.22.2
ipywidgets==8.1.2
ismember==1.0.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
MarkupSafe==2.0.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.14.3
opt-einsum==3.3.0
packaging==24.0
pandas==1.5.3
parso==0.8.3
patsy==0.5.6
pgmpy==0.1.25
pillow==10.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
pypickle==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-louvain==0.16
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
skrebate==0.62
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
statsmodels==0.14.1
sympy==1.12
tabulate==0.9.0
texttable==1.7.0
threadpoolctl==3.4.0
torch==2.2.2
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
widgetsnbextension==4.0.10
